pyproject.toml
[project]
name = "esp3-agent-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.123.10",
    "httpx>=0.28.1",
    "load-dotenv>=0.1.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "uvicorn>=0.38.0",
    "motor>=3.6.0",
    "python-multipart>=0.0.12",
    "mcp>=1.0.0",
    "dnspython>=2.6.1",
    "certifi>=2024.0.0",
]
